Add Code for Haxe Version
This commit contains all the necessary changes to migrate Skies Below to
the HaxeFlixel game engine.  There are no functional changes: the behavior
of the game itself should be indistinguishable from the Flash version.

In addition to necessary changes, this commit includes various
housekeeping tasks as well: unused code and files are removed, and
many comments are updated to be clearer and more accurate.